Moving back to the "source" element
Attempting to lump in the source element with the clones
was a bad idea, since the clones are not dynamicForms
(and even if they were, they'd have a completely
different copy of the plugin, so message passing would be
out). Having a centralized and standard launching-off
point is not that bad, I do wish this was more
generalized though.