Altering reflowFormNames to work on all elements
The benefit here is that any element with a name
property can now become a containing structure for the
resultant form data. This does impose some restrictions,
namely requiring you to repurpose the "name" attribute.
A better solution in the future might be to just use
dynamicFormName for everything (except the actual form
controls).